+++ properties.config
@@ -2,14 +2,14 @@
 Header=True
 Trailer=True
 separator=|
-OutputFileFormat=CSV
-DefaultHeader=False
-DefaultTrailer=False
-DefaultHeaderPath=//Users//shreyas//Documents//Eligibility_Automation//Header.csv
-DefaultTrailerPath=//Users//shreyas//Documents//Eligibility_Automation//Trailer.csv
-RecordDeclaration=//Users//shreyas//Documents//Eligibility_Automation//RecordDeclaration.csv
-OutputFilePath=//Users//shreyas//Documents//Eligibility_Automation//TestData
+OutputFileFormat=Text
+SeparateHeader=False
+SeparateTrailer=False
+DefaultHeaderPath=/Users/shreyas/Documents/Eligibility_Automation/Davita/Header
+DefaultTrailerPath=/Users/shreyas/Documents/Eligibility_Automation/Davita/Trailer
+RecordDeclaration=/Users/shreyas/Documents/Eligibility_Automation/Davita/DavitaRecordDeclaration.csv
+OutputFilePath=/Users/shreyas/Documents/Eligibility_Automation/Davita/TestData
 SSNFieldNumber=4
 SubscriberSSNFieldNumber=2
 RelationshipCodeNumber=9
-
+EmployeeFieldValue=01
